requirements.txt
boto3==1.26.157
botocore==1.29.157
PyYAML==6.0
black==23.3.0
isort==5.12.0
pylint==2.17.4
